Add work on move to seperate auth class for daemon
